Separate FingerprintingStrategy interface
