+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/TaskExecutionServices.java
@@ -36,12 +36,12 @@ import org.gradle.execution.plan.PlanExecutor;
 import org.gradle.initialization.BuildCancellationToken;
 import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.concurrent.ParallelismConfigurationManager;
+import org.gradle.internal.fingerprint.FingerprintCompareStrategy;
 import org.gradle.internal.fingerprint.HistoricalFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.impl.AbsolutePathFingerprintCompareStrategy;
 import org.gradle.internal.fingerprint.impl.ClasspathCompareStrategy;
 import org.gradle.internal.fingerprint.impl.DefaultHistoricalFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.impl.EmptyHistoricalFileCollectionFingerprint;
-import org.gradle.internal.fingerprint.impl.FingerprintCompareStrategy;
 import org.gradle.internal.fingerprint.impl.IgnoredPathCompareStrategy;
 import org.gradle.internal.fingerprint.impl.NormalizedPathFingerprintCompareStrategy;
 import org.gradle.internal.resources.ResourceLockCoordinationService;
